+++ o11nplugin-crypto-core/src/main/java/com/vmware/o11n/plugin/crypto/model/CryptoEncryption.java
@@ -16,6 +16,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.vmware.o11n.plugin.crypto.service.CryptoEncryptionService;
 import com.vmware.o11n.plugin.sdk.annotation.VsoMethod;
 import com.vmware.o11n.plugin.sdk.annotation.VsoObject;
 import com.vmware.o11n.plugin.sdk.annotation.VsoParam;
