+++ o11nplugin-crypto-core/src/main/java/com/vmware/o11n/plugin/crypto/service/CryptoCertificateService.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2017 VMware, Inc. All Rights Reserved.
  * SPDX-License-Identifier: BSD-2-Clause
  */
-package com.vmware.o11n.plugin.crypto.model;
+package com.vmware.o11n.plugin.crypto.service;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -38,6 +38,8 @@ import org.apache.commons.codec.digest.DigestUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
+
+import com.vmware.o11n.plugin.crypto.model.CryptoUtil;
 
 @Component
 public class CryptoCertificateService {
